Move fromBase64PrivateKey into OasisKey
